Move catalog group into its own file
